Refactor JSON array types on profile to DRY up implementation The current approach involves quite a bit of boilerplate for each JSONB attribute; I want to use a similar JSONB storage technique for allocation complex case answers, but without this boilerplate overhead. Introduce some generic classes that can be used for this moving forwards.
