+++ app/models/profile/profile_identifiers.rb
@@ -1,22 +1,13 @@
 # frozen_string_literal: true
 
 class Profile
-  class ProfileIdentifiers
-    extend Forwardable
-
-    def_delegators :@collection, *[].public_methods - %i[object_id __send__]
-
-    def initialize(array = [])
-      array = JSON.parse(array) if array.is_a? String
-      collection = Array(array).map do |item|
-        item.is_a?(Profile::ProfileIdentifier) ? item : Profile::ProfileIdentifier.new(item)
-      end
-
-      @collection = collection.reject(&:empty?)
+  class ProfileIdentifiers < Types::BaseCollection
+    def concrete_class
+      ProfileIdentifier
     end
 
-    def to_a
-      @collection
+    def remove_empty_items?
+      true
     end
   end
 end
